Remove unneded type tags for scripts
In HTML5 we can omit the type tag for scripts because it's useless in
javascript.

http://www.w3.org/html/wg/drafts/html/master/scripting-1.html#the-script
-element

"When used to include dynamic scripts, the scripts may either be
embedded inline or may be imported from an external file using the src
attribute. If the language is not that described by "text/javascript",
then the type attribute must be present, as described below. Whatever
language is used, the contents of the script element must conform with
the requirements of that language's specification."