Few changes to the Ubuntu Installation Guide
Changes: 

Removed `cd/path/to/nodebb/install` as we haven't instructed how to create the directory before changing to it. The git pull will create the nodebb directory in their root/profile folder anyway, which is fine for most installs. 

Added sudo to the base stack install.

Added command for checking Node.js version. 

The guide at the minute might seem a bit daunting for newer users, so if psychobunny doesn't mind, I could create a post over on the community for an up to date noobs guide on how to install using digital ocean as an example. Then if it helps, add it to the docs. (Also, changes in this github aren't showing up in docs.nodebb, the update to sudo apt-get update && sudo apt-get dist-upgrade aren't displayed.

Hope this helps.