Change text colour when searching custom CSS or JS
Not sure if this is the right place for this, but it seemed to be where the inherit rule was coming from. If you search for some text in the custom CSS or JS fields with the new IDE, the text appears the same colour as the input field. Needed more blaaaaack.