Add 'action:request.earlyProcessing' hook
Add a new hook in order to allow early request modifications. Will be useful in case if someone wants to modify request from inside a plugin before it will be processed by any other middleware. (For example it can be used for programmatic login, so `req.user` will be exposed to other middlewares)