refactor: navigator up/down to navigate by post index, #5321
The original behaviour of the up/down buttons in the navigator
approximated the page up and page down keys on the desktop. It was
determined (by me, mostly) that this was not useful, and a better
behaviour would be to update the selected index instead.

While updating this behaviour, I ran into a number of buggy code
that needed fixing:

- Off-by-one errors since server referred to post indices as an
  array starting at 0, while the client referred to them as an
  array starting at 1
- Code that caused scroll behaviour to continue adding more and
  more handlers the more you interacted with the navigator (that
  was a fun one to debug...)