Docker: Allow file based declaration of active modules
By default, the file is empty but a user can mount a different file into the container.

In K8s it is unsafe to first run the container and then later on activate modules. This should be done before nodebb is started.
Activating modules for each pod is safe as it is an idempotent operation.