feat: allow FormData object to be passed in to the API module
Currently, only objects can be passed in, and it is automatically serialized into json and sent via jQuery .ajax().

This PR extends the module so a FormData object can be passed in, and updates the module so it uses Fetch API instead of jQuery.

At this time regular requests continue to use jQuery for backwards compatibility.

Use case: file uploads via API.