fix: sanitize-html configuration passed in src/posts/parse.js
Cursory review of sanitize-html documentation suggests that the currently-used `globalAttributes` property no longer exists, but was replaced with `nonBooleanAttributes`, likely because the attribute allow-list explicitly applies only to "non-boolean" attributes (e.g. not `checked` or `selected`).

Either way it does not likely affect us but is mainly here for future-proofing purposes.