fix: update note assertion topic members check to simpler posts.exists check
The original logic checked that each member of the resolved chain was part of the resolved topic.

That isn't always the case, especially when topics splinter due to network timeouts/unavailability.

This ended up causing issues where already asserted posts were re-asserted but failed because they no longer served an _activitypub object since it was already asserted and the data was just pulled from the db.