ci: drop ARM v7 from docker builds
Removed optional ARM v7 (32 bit) platform from the workflow due to lack of support from Node and very limited usefulness.

As the platform had been flaky in the past, this part of the workflow was already optional and didn't cause it to fail. So this is just the next step here.